+++ datanode/networkhistory/service_test.go
@@ -379,12 +379,12 @@ func TestMain(t *testing.M) {
 		log.Infof("%s", goldenSourceHistorySegment[4000].HistorySegmentID)
 		log.Infof("%s", goldenSourceHistorySegment[5000].HistorySegmentID)
 
-		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[1000].HistorySegmentID, "QmQEkraFHXm2wqben3U1JsVBntLTakHZm7q3uK1T1gehiS", snapshots)
-		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[2000].HistorySegmentID, "QmRrsekJy3kSwpnriZsyrhmCy5uwqhUuV5bPYpg8AAoqNo", snapshots)
-		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[2500].HistorySegmentID, "QmYUrbLT58tsiEdGzCqfFybVpqLuRmfazJbpsoLQHxcRx5", snapshots)
-		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[3000].HistorySegmentID, "QmUx3Y8HdstFDwYbhq2LV63tKMg5V1DyRj2CQz24ZYwowY", snapshots)
-		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[4000].HistorySegmentID, "QmaxSz1ZYrGNRUEP8ihTfwTo2JePZUMjKHVftnYxcHPHnX", snapshots)
-		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[5000].HistorySegmentID, "QmfRtpNVPbGjsiBHBPkKjKV46iCxYNDhxgyHYPr11TGsQk", snapshots)
+		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[1000].HistorySegmentID, "QmadsG2uSkvh9h1qgEF7HzUWMGeTrLDG88fTLCykRhvPiG", snapshots)
+		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[2000].HistorySegmentID, "QmSVi1Tw8eHgt99wzydzyHs8WZDD84eXUioHoPJ2TUo6eL", snapshots)
+		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[2500].HistorySegmentID, "QmZ3DPxJk9GwMDK7KPpMsNXKy8jVj3AmdGRvSWv4Ee14U4", snapshots)
+		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[3000].HistorySegmentID, "QmVipESCMGuJtyYZY7FvPaMgWP5gBFG9o9ncN1U9dQR4s9", snapshots)
+		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[4000].HistorySegmentID, "QmSyrLAoj6BHRkNK44oqeUXNdsVwbuwAVrjyC9Y19GpMch", snapshots)
+		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[5000].HistorySegmentID, "Qme9eMuuukaA16swo8mxFpuXuF2RsVUYQ2GtELqNdHuSSa", snapshots)
 	}, postgresRuntimePath, sqlFs)
 
 	if exitCode != 0 {