+++ CHANGELOG.md
@@ -51,6 +51,7 @@
 - [10000](https://github.com/vegaprotocol/vega/issues/10000) - Introduce `cometbtf` command to replace `tendermint`.
 - [9948](https://github.com/vegaprotocol/vega/issues/9948) - Add support for linked stop orders.
 - [9849](https://github.com/vegaprotocol/vega/issues/9849) - Add database support for `num.Uint`.
+- [10275](https://github.com/vegaprotocol/vega/issues/10275) - Add API to list party's margin mode.
 
 ### 🐛 Fixes
 
@@ -101,6 +102,7 @@
 - [10276](https://github.com/vegaprotocol/vega/issues/10276) - `ListCandleData` should return empty strings instead of zero for prices when being gap filled.
 - [10278](https://github.com/vegaprotocol/vega/issues/10278) - Transfers connection filtering by `isReward` and party causes error
 - [10251](https://github.com/vegaprotocol/vega/issues/10251) - Add batch proposal API and fix batch proposal submission.
+- [10285](https://github.com/vegaprotocol/vega/issues/10285) - Fix `MTM` settlement panic.
 
 ## 0.73.0